refactor(sysAdmin): address self-review feedback
1. disabled={loading} の条件を修正 — cache-and-network +
   notifyOnNetworkStatusChange: true の組み合わせでバックグラウンド再取得中も
   loading=true になり、slider/filter が毎回 disabled になっていた。
   disabled={loading && !data} で初回ロード中のみに限定。
2. _shared/__mocks__/ を _shared/fixtures/ にリネーム — __mocks__ は
   Vitest/Jest のモジュール自動モック予約名であり、テスト/Story 用の
   フィクスチャファクトリを置く場所として誤解を招くため。
3. useMemberListPagination を ref ベースに — loadMore の identity を空 deps
   の useCallback で真に安定化。react-window に渡すハンドラの再生成を抑制。
4. auto-load useEffect に無限ループ防止ガードを追加 — hasNextPage=true が
   返り続けるのに users が増えないケースで lastAutoLoadLengthRef と比較し
   以後の auto-load を停止。
