+++ src/app/sysAdmin/features/communityDetail/components/CommunityDashboardDetail.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import { GetSysAdminCommunityDetailDocument } from "@/types/graphql";
 import { withApollo } from "../../../../../../.storybook/decorators";
 import { CommunityDashboardDetail } from "./CommunityDashboardDetail";
-import { makeCommunityDetailPayload } from "../../../_shared/__mocks__/sysAdminDashboard";
+import { makeCommunityDetailPayload } from "../../../_shared/fixtures/sysAdminDashboard";
 
 const meta: Meta<typeof CommunityDashboardDetail> = {
   title: "SysAdmin/Detail/CommunityDashboardDetail",
