+++ src/app/sysAdmin/features/dashboard/components/PlatformSummaryCards.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import { PlatformSummaryCards } from "./PlatformSummaryCards";
-import { makePlatformSummary } from "../../../_shared/__mocks__/sysAdminDashboard";
+import { makePlatformSummary } from "../../../_shared/fixtures/sysAdminDashboard";
 
 const meta: Meta<typeof PlatformSummaryCards> = {
   title: "SysAdmin/Dashboard/PlatformSummaryCards",
